start with candles chart and separating everything into components for easier reading
