+++ src/components/starter_example.vue
@@ -1,0 +1,22 @@
+<template>
+  <div id="example">
+    example
+  </div>
+</template>
+
+<script>
+export default {
+  data () {
+    return {
+    }
+  },
+  methods: {
+  },
+  mounted () {
+  },
+  name: 'example'
+}
+</script>
+
+<style scoped>
+</style>
